requirements.txt
beautifulsoup4==4.7.1
Django==2.2.2
django-admin-numeric-filter==0.1.5
django-admin-rangefilter==0.5.0
django-pandas==0.6.1
djangorestframework==3.10.2
finance-datareader==0.7.2
Keras==2.2.4
numpy==1.16.4
pandas==0.24.2
pykrx==0.1.7
requests==2.22.0
TA-Lib==0.4.17


